requirements.txt
Flask==2.3.3
supabase==2.0.2
python-dotenv==1.0.0
psycopg2-binary==2.9.10
Werkzeug==2.3.7 